// src/app/state/movies.enum.ts
export enum MoviesActionEnum {
  GET_MOVIES = '[Movies Page] Get Movies',
  GET_MOVIES_SUCCESS = '[Movies Page] Get Movies Success',
  GET_MOVIES_FAILED = '[Movies Page] Get Movies Failed',
  ADD_MOVIE = '[Movies Page] Add Movies',
  ADD_MOVIE_SUCCESS = '[Movies Page] Add Movies Success',
  ADD_MOVIE_FAILED = '[Movies Page] Add Movies Failed',
  EDIT_MOVIE = '[Movies Page] Edit Movies',
  EDIT_MOVIE_SUCCESS = '[Movies Page] Edit Movies Success',
  EDIT_MOVIE_FAILED = '[Movies Page] Edit Movies Failed',
  DELETE_MOVIE = '[Movies Page] Delete Movies',
  DELETE_MOVIE_SUCCESS = '[Movies Page] Delete Movies Success',
  DELETE_MOVIE_FAILED = '[Movies Page] Delete Movies Failed',
}